+++ common/src/main/java/net/frozenblock/lib/entity/impl/variant/ConfigCheck.java
@@ -1,0 +1,23 @@
+package net.frozenblock.lib.entity.impl.variant;
+
+import com.mojang.serialization.MapCodec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.frozenblock.lib.config.v2.entry.predicates.ConfigPredicate;
+import net.minecraft.world.entity.variant.SpawnCondition;
+import net.minecraft.world.entity.variant.SpawnContext;
+
+public record ConfigCheck(ConfigPredicate configPredicate) implements SpawnCondition {
+	public static final MapCodec<ConfigCheck> MAP_CODEC = RecordCodecBuilder.mapCodec(instance -> instance.group(
+		ConfigPredicate.CODEC.fieldOf("predicate").forGetter(ConfigCheck::configPredicate)
+	).apply(instance, ConfigCheck::new));
+
+	@Override
+	public MapCodec<? extends SpawnCondition> codec() {
+		return MAP_CODEC;
+	}
+
+	@Override
+	public boolean test(SpawnContext spawnContext) {
+		return this.configPredicate.test();
+	}
+}